backend/requirements.txt
# Backend API - install with: pip install -r requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
alembic>=1.13.0
python-dotenv>=1.0.0
# Optional: for real ML explanations (load simple_model.pkl and use feature importance)
scikit-learn>=1.0.0
# Share graphic generation
Pillow>=10.0.0
# Stripe payments
stripe>=8.0.0
